linkedin_analyzer/requirements.txt
Flask==3.0.2
werkzeug==3.0.1
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.29
pypdf==4.1.0
python-dotenv==1.0.1
